[base] add network_data into the connection
connection structure contains read and write network callbacks. This
callbacks may be changed by a module like mod_openssl or mod_mbedtls.
But only one module may be change them and the callback is not able to
retrieve the module information directly.
Mod_openssl use the system of singleton like that:
    handler_ctx *hctx = con->plugin_ctx[plugin_data_singleton->id];

But if during the initialisation of the callbacks into CONNECTION_FUNC,
it is possible to set information about the plugin like that:
    plugin_data *p = p_d;
    con->network_read = connection_read_cq_ssl;
    con->network_write = connection_write_cq_ssl;
    con->network_data = p;

and the callbacks become:
    handler_ctx *hctx = con->plugin_ctx[((plugin_data *)con->network_data)->id];